# docs/openapi/presentation-api-v1.yaml
openapi: 3.1.0
info:
  title: IPO Screening Engine — Presentation Read-Model API
  version: 1.0.0
  description: |
    Deterministic, read-only presentation API projecting immutable evaluation,
    post-listing observation, backtesting, and calibration records produced by
    the Indian Mainboard IPO Screening Engine.

    PRINCIPLES:
    - Pure presentation read model (GET / HEAD / OPTIONS only; zero mutation endpoints).
    - Authoritative source of truth remains the immutable EvaluationStore on disk.
    - Zero client-side or API-side duplicate scoring or metric calculation.
    - Explicit preservation of UNKNOWN, NOT_APPLICABLE, and INSUFFICIENT_DATA semantics.
    - Zero exposure of model provider or external data credentials.
servers:
  - url: /api/v1
    description: Presentation API v1 prefix
paths:
...
  /ipos:
    get:
      summary: Discover and list evaluated IPO issuers
      operationId: listIpos
      tags: [IPOs]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Substring match against company_name or ipo_id
          schema:
            type: string
        - name: sector_profile
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of IPO issuers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpoListResponse'
  /ipos/{ipo_id}:
    get:
      summary: Retrieve an IPO issuer and its evaluation summary
      operationId: getIpo
      tags: [IPOs]
      parameters:
        - name: ipo_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authoritative IPO details and evaluation timeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpoDetail'
        '404':
          description: IPO identifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /ipos/{ipo_id}/history:
    get:
      summary: Retrieve chronological evaluation history and lifecycle deltas
      operationId: getIpoHistory
      tags: [IPOs]
      parameters:
        - name: ipo_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chronological list of evaluations with preliminary-to-final delta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpoHistoryResponse'
        '404':
          description: IPO identifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /evaluations:
    get:
      summary: Discover evaluations across issuers with deterministic filtering
      operationId: listEvaluations
      tags: [Evaluations]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: ipo_id
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [PRELIMINARY, FINAL]
        - name: verdict
          in: query
          schema:
            type: string
            enum: [APPLY, CONSIDER, AVOID, INSUFFICIENT_DATA]
        - name: confidence
          in: query
          schema:
            type: string
            enum: [High, Medium, Low]
        - name: engine_version
          in: query
          schema:
            type: string
        - name: config_version
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated evaluations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationListResponse'
  /evaluations/{evaluation_id}:
    get:
      summary: Retrieve complete evaluation scorecard detail
      operationId: getEvaluation
      tags: [Evaluations]
      parameters:
        - name: evaluation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Complete immutable evaluation scorecard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationDetail'
        '404':
          description: Evaluation identifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /evaluations/{evaluation_id}/evidence:
    get:
      summary: Retrieve evidence registry and citations for an evaluation
      operationId: getEvaluationEvidence
      tags: [Evidence]
      parameters:
        - name: evaluation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Evidence citations and verbatim prospectus quotes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
        '404':
          description: Evaluation identifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /evidence/{evidence_id}:
    get:
      summary: Retrieve specific evidence citation by evidence identifier
      operationId: getEvidenceItem
      tags: [Evidence]
      parameters:
        - name: evidence_id
          in: path
          required: true
          description: Evidence identifier or compound identifier {eval_id}:{evidence_id}
          schema:
            type: string
      responses:
        '200':
          description: Evidence item details with provenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceDetail'
        '404':
          description: Evidence item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /evaluations/{evaluation_id}/post-listing:
    get:
      summary: Retrieve post-listing price observations and corporate actions
      operationId: getPostListingObservations
      tags: [Post-Listing]
      parameters:
        - name: evaluation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Multi-horizon post-listing price and return observations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListingObservationsResponse'
        '404':
          description: Evaluation identifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /evaluations/{evaluation_id}/performance:
    get:
      summary: Retrieve realized returns and benchmark alpha summary
      operationId: getPerformanceSummary
      tags: [Post-Listing]
      parameters:
        - name: evaluation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Structured horizon returns and Nifty 50 alpha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSummaryResponse'
        '404':
          description: Evaluation identifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /backtest/analytics:
    get:
      summary: Retrieve latest backtest analytics and statistical diagnostics
      operationId: getBacktestAnalytics
      tags: [Backtest]
      responses:
        '200':
          description: Backtest statistical diagnostics (Rank IC, hit rates, deciles)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestAnalyticsResponse'
        '404':
          description: Backtest analysis report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /calibration/proposals/{proposal_id}:
    get:
      summary: Retrieve detailed calibration proposal artifact
      operationId: getCalibrationProposal
      tags: [Calibration]
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full calibration proposal artifact with shadow evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrationProposal'
        '404':
          description: Calibration proposal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /ingest/document:
    post:
      summary: Ingest and evaluate an IPO filing document (DRHP/RHP PDF)
      description: |
        Uploads a PDF filing document, extracts domain entities using DocumentExtractor,
        converts into CanonicalInput, validates, runs deterministic evaluation,
        and persists the record to the immutable EvaluationStore.
        Duplicate uploads return the existing evaluation idempotently.
      operationId: ingestDocument
      tags: [Ingestion]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF filing document (DRHP/RHP)
                mode:
                  type: string
                  enum: [final, preliminary]
                  default: final
                  description: Evaluation mode
      responses:
        '200':
          description: Evaluation already exists (duplicate idempotent return)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResponse'
        '201':
          description: Document successfully ingested and new evaluation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResponse'
        '400':
          description: Bad request (invalid PDF magic bytes, empty file, or invalid mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Payload too large (file exceeds 50 MB limit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Document extraction or evaluation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
...
    IpoDetail:
      type: object
      required: [ipo_id, company_name, evaluations]
      properties:
        ipo_id:
          type: string
        company_name:
          type: string
        sector_profile:
          type: string
          nullable: true
        icdr_route:
          type: string
          nullable: true
        evaluations:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationSummary'
    IpoListResponse:
      type: object
      required: [items, total, page, page_size, pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IpoSummary'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        pages:
          type: integer
...
    EvaluationListResponse:
      type: object
      required: [items, total, page, page_size, pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationSummary'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        pages:
          type: integer
...
    EvaluationDetail:
      type: object
      required: [evaluation_id, ipo_id, company_name, evaluation_mode, evaluation_timestamp, engine_version, spec_version, config_version, config_hash, input_snapshot_hash, source_manifest_hash, result_hash, score, knockouts, verdict, penalties, missing_unverified, provenance]
      properties:
        evaluation_id:
          type: string
        ipo_id:
          type: string
        company_name:
          type: string
        evaluation_mode:
          type: string
          enum: [PRELIMINARY, FINAL]
        evaluation_timestamp:
          type: string
          format: date-time
        engine_version:
          type: string
        spec_version:
          type: string
        config_version:
          type: string
        config_hash:
          type: string
        input_snapshot_hash:
          type: string
        source_manifest_hash:
          type: string
        market_snapshot_hash:
          type: string
          nullable: true
        peer_snapshot_hash:
          type: string
          nullable: true
        result_hash:
          type: string
        score:
          $ref: '#/components/schemas/EvaluationScoreDetail'
        knockouts:
          $ref: '#/components/schemas/KnockoutSummary'
        penalties:
          type: array
          items:
            $ref: '#/components/schemas/PenaltyItem'
        verdict:
          type: object
          properties:
            verdict:
              type: string
            band_score:
              type: string
            uncertain:
              type: boolean
            insufficient_data:
              type: boolean
            mode_label:
              type: string
              nullable: true
        missing_unverified:
          type: array
          items:
            $ref: '#/components/schemas/MissingUnverifiedItem'
        preliminary_delta:
          type: object
          nullable: true
        provenance:
          type: object
          properties:
            sources:
              type: array
              items:
                type: string
            evidence_count:
              type: integer
            engine_version:
              type: string
            spec_version:
              type: string
    EvidenceDetail:
      type: object
      required: [evidence_id, field]
      properties:
        evidence_id:
          type: string
        field:
          type: string
        source:
          type: string
          nullable: true
        document:
          type: string
          nullable: true
        page:
          type: integer
          nullable: true
        locator:
          type: string
          nullable: true
        quote:
          type: string
          nullable: true
        extracted_value:
          type: object
          nullable: true
        unit:
          type: string
          nullable: true
    EvidenceResponse:
      type: object
      required: [evaluation_id, source_manifest_hash, evidence_count, items]
      properties:
        evaluation_id:
          type: string
        source_manifest_hash:
          type: string
        sources:
          type: array
          items:
            type: string
        evidence_count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceDetail'
    IpoHistoryResponse:
      type: object
      required: [ipo_id, company_name, history]
      properties:
        ipo_id:
          type: string
        company_name:
          type: string
        history:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationSummary'
        delta:
          type: object
          nullable: true
...
    PostListingObservationsResponse:
      type: object
      required: [final_evaluation_id, ipo_id, observation_count, observations]
      properties:
        final_evaluation_id:
          type: string
        ipo_id:
          type: string
        observation_count:
          type: integer
        observations:
          type: array
          items:
            $ref: '#/components/schemas/PostListingObservationItem'
...
    PerformanceSummaryResponse:
      type: object
      required: [final_evaluation_id, ipo_id, company_name, horizons]
      properties:
        final_evaluation_id:
          type: string
        ipo_id:
          type: string
        company_name:
          type: string
        issue_price:
          type: string
          nullable: true
        listing_date:
          type: string
          nullable: true
        horizons:
          type: object
          properties:
            listing_day:
              $ref: '#/components/schemas/HorizonPerformance'
            one_week:
              $ref: '#/components/schemas/HorizonPerformance'
            one_month:
              $ref: '#/components/schemas/HorizonPerformance'
            six_month:
              $ref: '#/components/schemas/HorizonPerformance'
...
    ApiError:
      type: object
      required: [error, message, code]
      properties:
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Evaluation 01ba66c... not found in store
        code:
          type: string
          example: EVALUATION_NOT_FOUND
        details:
          type: object
          nullable: true
...